Add descriptive comments

+++ orchest/memory-store/app/manager.py
@@ -9,7 +9,8 @@ def construct_pipeline(pipeline_fname='pipeline.json'):
     with open(pipeline_fname, 'r') as f:
         description = json.load(f)
 
-    pipeline = nx.DiGraph()
+    auto_eviction = description['settings'].get('auto-eviction', False)
+    pipeline = nx.DiGraph(auto_eviction=auto_eviction)
 
     # Create and add nodes.
     steps = description['steps']
@@ -73,58 +74,77 @@ def delete(client, uuids):
 
 
 def start_manager(store_socket_name, pipeline_fname):
+    # Connect to the plasma store and subscribe to its notification
+    # socket.
     client = plasma.connect(store_socket_name)
-
     client.subscribe()
 
-    # This one should actually be kept in memory
+    # Keeps a `pipeline` in memory to maintain state. Everytime a step
+    # retrieves the output from another the weight of that connection
+    # (aka edge) is set to 1. If the outdegree of a step is equal to the
+    # sum of the weight of its outgoing edges, then we know that all the
+    # receiving steps have already read the output. If the
+    # `auto-eviction` is set in the `pipeline.json`, then this will
+    # cause that output to be removed from the store.
     pipeline = construct_pipeline(pipeline_fname=pipeline_fname)
 
     while True:
         try:
             obj_id, data_size, mdata_size = client.get_next_notification()
         except OSError:
-            # "OSError: Failed to read object notification from Plasma socket"
+            # "Failed to read object notification from Plasma socket"
             continue
 
-        print(obj_id)
-
         mdata = client.get_metadata([obj_id], timeout_ms=1000)
-        print('mdata:', mdata)
 
-        # This is the case when a sealed object is deleted. Then also
-        # a notification is triggered. But we don't want to inspect
-        # this one.
+        # Whenever a sealed object is deleted, it also triggers a
+        # notification. However, we do not need to check for eviction in
+        # that case.
         if mdata[0] is None:
             continue
 
+        # TODO: change print to logging to stdout
+        print('Received:', obj_id)
         mdata = bytes(mdata[0])
 
-        # TODO: we should add some information about what it in the
-        # metadata. Currently it can be all sorts of strings. This is
-        # super hard to resolve.
-        identifier, _, data = mdata.partition(b';')
-        print(identifier, data)
+        # An example message: b'2;uuid-1,uuid-2'. Meaning that step with
+        # 'uuid-2' has retrieved the output from step with 'uuid-1'.
+        identifier, _, mdata = mdata.partition(b';')
         if identifier != b'2':
             continue
 
-        decoded_mdata = data.decode(encoding='utf-8')
+        decoded_mdata = mdata.decode(encoding='utf-8')
         source, target = decoded_mdata.split(',')
 
-        # Create new pipeline and propagate weights.
+        # Create new pipeline and propagate weights. A new pipeline is
+        # created to account for a possible change in the pipeline. A
+        # user might have added or removed multiple steps or
+        # connections.
         new_pipeline = construct_pipeline(pipeline_fname=pipeline_fname)
         propagate_weights(pipeline, new_pipeline)
+
+        # The `pipeline` instance is kept in memory so that the state of
+        # the eviction manager is maintained during the lifecycle of the
+        # store.
         pipeline = new_pipeline
 
         # Set that the target uuid has received from the source.
         pipeline[source][target]['weight'] = 1
 
-        # TODO: ONLY DELETE IF AUTO_EVICT OPTION IS SET IN PIPELINE.JSON
-        # Delete the uuids to evict.
+        # TODO: should we check for this options earlier, because
+        #       probably we want to start counting the moment the user
+        #       selects the options (and by deselect maybe reset all
+        #       weights to zero).
+        # Only consider evicting objects if the option is set.
+        if not pipeline.graph.get('auto_eviction', False):
+            continue
+
+        # Delete the objects corresponding to the `uuids_to_evict`.
         uuids_to_evict = get_uuids_to_evict(new_pipeline)
-        print(uuids_to_evict)
-        print([type(uuid) for uuid in uuids_to_evict])
         delete(client, uuids_to_evict)
 
-        # Need to also delete the object itself that contained the metadata.
+        print('Evicting:', uuids_to_evict)
+
+        # Need to also delete the "ping" object that contained the
+        # metadata.
         client.delete([obj_id])
